Apply PCA to all generate geometries
Use PCA to all geometries (not just planar ones) to align markers to a
'canonical' basis. This way, x-axis will be longest dimension of tool,
and y-axis second-largest. z-axis will complete orthonormal basis for
non-planar geometries, or be zero for planar ones.

Also adds warning when planar flag is enabled for non-planar tools, and
suggestion to enable planar flag in opposite case.