Tool maker utility (#4)
* Add isocenter calculations

* Reduce dimension via PCA for planar geometries

For geometries that user has indicated are planar, use PCA to eliminate
last principal component i.e. orthogonal direction of smallest variance
of markers. Essentially chooses plane that lies in directions of largest
variance of the markers, and projects markers on to it.

* Apply PCA to all generate geometries

Use PCA to all geometries (not just planar ones) to align markers to a
'canonical' basis. This way, x-axis will be longest dimension of tool,
and y-axis second-largest. z-axis will complete orthonormal basis for
non-planar geometries, or be zero for planar ones.

Also adds warning when planar flag is enabled for non-planar tools, and
suggestion to enable planar flag in opposite case.

* Check reference correspondence is one-to-one

When ordering markers to match the order of the reference
makers, make sure we find a one-to-one correspondence

* Refactor tool_maker script

Restructure to make control flow more explicit and do away with global
shared state

* Choose positive basis orientation

When aligning data to new basis using PCA, choose basis to have positive
orientation

Also sort JSON keys when dumping to file to make it easier to read

* Order and filter records as they are collected

Rather than collect all records and then order and filter them, we order
and filter as they are received from the ROS topic. This way sample
count is accurate, and code is a little cleaner.

* Add optional ID to tool geometry files

* Simplify marker re-ordering

* Add unit conversion to tool_maker script