Retrieve device option list during init
User manual states option ids associated with each device option name
may change between SDK versions, so option list is queried. When
attempting to set an option, we now first check that it is supported,
and for integer-value options also check whether it is within the valid
range for that option.